tf-idf filter seperated from keyword search
